Feature/RiverLane and PositionRefactor

+++ src/main/java/feup2526/ldts/t02g03/model/MovableLane.java
@@ -2,9 +2,9 @@ package feup2526.ldts.t02g03.model;
 
 public abstract class MovableLane extends Lane {
     protected Direction direction;
-    protected int speed;
+    protected double speed;
 
-    public MovableLane(int row, Direction direction, int speed) {
+    public MovableLane(int row, Direction direction, double speed) {
         super(row);
         if (direction == null) throw new IllegalArgumentException("Lane requires direction");
         if (direction != Direction.LEFT && direction != Direction.RIGHT) throw new IllegalArgumentException("Lane only supports LEFT/RIGHT directions");
@@ -16,5 +16,5 @@ public abstract class MovableLane extends Lane {
 
     public Direction getDirection(){return direction;}
 
-    public int getSpeed(){return speed;}
+    public double getSpeed(){return speed;}
 }
