+++ src/main/java/feup2526/ldts/t02g03/model/Position.java
@@ -1,20 +1,20 @@
 package feup2526.ldts.t02g03.model;
 
 public final class Position {
-    private int x,y;
+    private double x,y;
 
-    public Position(int x, int y) {
+    public Position(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {return x;}
-    public int getY() {return y;}
+    public double getX() {return x;}
+    public double getY() {return y;}
 
-    public Position up(){return new Position(x,y-1);}
-    public Position down(){return new Position(x,y+1);}
-    public Position left(){return new Position(x-1,y);}
-    public Position right(){return new Position(x+1,y);}
+    public Position up(){return new Position(x,y-1.0);}
+    public Position down(){return new Position(x,y+1.0);}
+    public Position left(){return new Position(x-1.0,y);}
+    public Position right(){return new Position(x+1.0,y);}
     public Position translate(int dx, int dy) {
         return new Position(x+dx,y+dy);
     }
